hsapp: move setupFonts tinto setup.go; a.setup is now called by a.Run (not by setupFonts)
